move trustroot resync period configration to different package
