+++ pkg/reconciler/trustroot/controller.go
@@ -16,7 +16,6 @@ package trustroot
 
 import (
 	"context"
-	"time"
 
 	"k8s.io/client-go/tools/cache"
 	kubeclient "knative.dev/pkg/client/injection/kube/client"
@@ -30,14 +29,13 @@ import (
 	"github.com/sigstore/policy-controller/pkg/apis/config"
 	trustrootinformer "github.com/sigstore/policy-controller/pkg/client/injection/informers/policy/v1alpha1/trustroot"
 	trustrootreconciler "github.com/sigstore/policy-controller/pkg/client/injection/reconciler/policy/v1alpha1/trustroot"
+	"github.com/sigstore/policy-controller/pkg/tuf"
 	cminformer "knative.dev/pkg/injection/clients/namespacedkube/informers/core/v1/configmap"
 )
 
 // This is what the default finalizer name is, but make it explicit so we can
 // use it in tests as well.
 const FinalizerName = "trustroots.policy.sigstore.dev"
-
-type trustrootResyncPeriodKey struct{}
 
 // NewController creates a Reconciler and returns the result of NewImpl.
 func NewController(
@@ -78,22 +76,8 @@ func NewController(
 			pkgreconciler.NamespaceFilterFunc(system.Namespace()),
 			pkgreconciler.NameFilterFunc(config.SigstoreKeysConfigName)),
 		Handler: controller.HandleAll(grCb),
-	}, FromContextOrDefaults(ctx)); err != nil {
+	}, tuf.FromContextOrDefaults(ctx)); err != nil {
 		logging.FromContext(ctx).Warnf("Failed configMapInformer AddEventHandlerWithResyncPeriod() %v", err)
 	}
 	return impl
-}
-
-func ToContext(ctx context.Context, duration time.Duration) context.Context {
-	return context.WithValue(ctx, trustrootResyncPeriodKey{}, duration)
-}
-
-// FromContextOrDefaults returns a stored trustrootResyncPeriod if attached.
-// If not found, it returns a default duration
-func FromContextOrDefaults(ctx context.Context) time.Duration {
-	x, ok := ctx.Value(trustrootResyncPeriodKey{}).(time.Duration)
-	if ok {
-		return x
-	}
-	return controller.DefaultResyncPeriod
 }
